+++ app/Werewolf/Command/Protect.hs
@@ -25,7 +25,7 @@ import Control.Monad.Writer
 import Data.Text (Text)
 
 import Game.Werewolf
-import Game.Werewolf.Command.Defender
+import Game.Werewolf.Command.Protector
 
 import Werewolf.Game
 import Werewolf.Messages